requirements.txt
python-stdnet=0.8.2
tornado=3.1.1